recapupload: date tz workaround
Don't report the date to CourtListener in UTC, convert to the local zone.

Fortunately (today) we only care about dates, not times, so this isn't so
serious. [But we should care about times!]

Unfortunately recapupload doesn't know what timezone each court is in,
so converting to local time is hard. But as a compromise, convert to
US/Pacific, which is less wrong (and a 10pm Pacific filing is much
more likely than a 1am Eastern filing).

Also, unfortunately, the only way to control the local zone is to set
the TZ environment variable and call tzset(). And then we have to put
it back in case there are others in the same process who use time
localization.