recapupload: Allow [] in parsing regexp
Turns out that there are square brackets [] inside item strings,
despite the fact that they are delimited by the outer set of
[]. Again, this maybe a condition not really encountered in the RSS
XML in the wild, but, e.g., [3] references are found in the wild as a
reference to document entry 3, and thus on RSS-format back-composed
from NEFs, they are present.

So just use a greedy regexp and suck up everything between the outer
brackets. That is:
  .*
instead of
  [^\]]*